created stamps
Define a procedure, stamps, which takes as its input a positive integer
in pence and returns the number of 5p, 2p and 1p stamps (p is pence)
required to make up that value. The return value should be a tuple of
three numbers (that is, your return statement should be followed by the
number of 5p, the number of 2p, and the nuber of 1p stamps).