Created Max_pages
The web crawler we built at the end of Unit 3 has some serious flaws if
we were going to use it in a real crawler. One problem is if we start
with a good seed page, it might run for an extremely long time (even
forever, since the number of URLS on the web is not actually finite).
This question and the following one explore two different ways to limit
the pages that it can crawl.