created improving the index
The current index includes a url in the list of urls
for a keyword multiple times if the keyword appears
on that page more than once.

It might be better to only include the same url
once in the url list for a keyword, even if it appears
many times.
Modify add_to_index so that a given url is only
included once in the url list for a keyword,
no matter how many times that keyword appears.